requirements.txt
pandas
numpy
nltk
scikit-learn
langchain
sentence-transformers
flask
fastapi
uvicorn[standard]
psycopg2-binary
streamlit
torchvision
langgraph
toml
requests
Pillow

mlflow>=2.8.0
mlflow[extras]>=2.8.0

prometheus-client>=0.17.0
psutil>=5.9.0

PyYAML>=6.0
python-dotenv>=0.19.0

gunicorn>=20.0.0

chromadb>=0.4.15
typing_extensions>=4.8.0

# PySpark for distributed computing
pyspark>=3.5.0

# Redis for Feature Store
redis>=5.0.0
hiredis>=2.2.0

